+++ lib/services/services.dart
@@ -8,5 +8,7 @@ export 'update_service.dart';
 export 'speech_service.dart';
 export 'dictionary_service.dart';
 export 'notification_service.dart';
+export 'guidebook_service.dart';
+